* Carry ISO 3166 in a 16-bit integer instead of a 32-bit integer. Due to    structure padding this may save a little bit of space in some places.

+++ src/ui/gtk/gtk2/nodes.c
@@ -155,7 +155,7 @@ struct node_data {
 	size_t info_size;
 	guint uptime;
 	guint connected;
-	gint country;
+	guint16 country;
 	GtkTreeIter iter;
 	gchar version[24];
 	gchar flags[16 + sizeof "<tt></tt>"];
